About the Project
Predicting Demand for a Bicycle Sharing Company.
Introduction
Bicycle Sharing Systems are a means of renting bicycles where the process of obtaining membership, rental,and bike return is automated via a network of bike-stations located throughout a city. The users of bike sharing systems can pick up bicycles from a bike-station in one location and return them to another in possibly any location of the city.Bike Sharing System ensures that pollution is reduced with use of bicycles there is reduction in use of motor vehicles which leads to reduction in emission of pollutants in the air.

Problem-Statement
In a bicycle sharing systems it is very important for the administrators to know how many cycles will be needed in each bicycle station, knowing this count, it enables them to arrange proper number of cycles at the stations and decide whether a particular station needs to have extra number of bicycle at that stands. Utilizing a seperate weather data and not the usage pattern here, its needed to forecast bike rental demand in the Capital Bikeshare program in Washington, D.C.

Data-Source
The dataset in this project is provided by Kaggle and is an open dataset hosted at UCI Machine Learning Repository

Project-Description
This project work, focuses on which algorithm can work better for the real world problem of bicycle sharing demand prediction and there after recommending which are the important factors for predicting the demand.